requirements.txt
astroid==3.3.6
bcrypt==5.0.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
dill==0.3.9
dominate==2.9.1
Faker==33.1.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-sqlalchemy-stubs==0.2
Flask-WTF==1.2.2
greenlet==3.1.1
humanize==4.11.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.13.0
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==24.2
platformdirs==4.3.6
Pygments==2.18.0
pylint==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.36
sqlalchemy-stubs==0.4
sqlalchemy2-stubs==0.0.2a38
tomlkit==0.13.2
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.2.1
